# MAIN_project/static/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'Maximum'
  version: ''
paths:
...
  /objects/:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get object's list by token
      operationId: Objects
      description: |
        # Получение списка предметов:
        ## Суперпользователь - абсолютно все предметы
        ## Активный пользователь - просмотр всех предметов в активных школах, в которых состоит и request.user
      parameters:
        - in: query
          name: school_id
          schema:
            type: integer
          description: filter by school_id
        - in: query
          name: search
          schema:
            type: string
          description: Search by name
        - in: query
          name: ordering
          schema:
            type: string
          description: Order by id, name, school
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
        - in: query
          name: offset
          schema:
            type: integer
          description: The number of items to skip before starting to collect the result set
      responses:
        '200':
          content:
            application/json:
              schema: { }
          description: OK
      tags:
        - objects

    post:
      security:
        - bearerAuth: [ ]
      operationId: CreateObject
      summary: Create a object
      description: |
        # Создание предмета
        ## Cуперпользователь - создание предметов для любых школ
        ## Активный завуч - создание предметов только в своей активной школе
        ## В request body перечислены все возможные поля (все обязательные)
        ### Обязательные поля: "name", "school_id"
        ### Проверка на валидацию school_id
        ### Проверка на уникальность названия внутри школы
      parameters: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: name
                school_id:
                  type: integer
                  example: 1
      responses:
        '201':
          content:
            application/json:
              schema: { }
          description: OK
      tags:
        - objects
...
  /schedules/:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get time schedule's list by token
      operationId: Schedules
      description: |
        # Получение расписания уроков:
        ## Суперпользователь - абсолютно все 
        ## Активный пользователь - просмотр всех данных в активных школах, в которых состоит и request.user
        ## Фильтрация:
        ### Начало урока - нижняя граница (не раньше чем)
        ### Конец урока - верхняя граница (не позже чем)
        ### Номер урока
        ### Отмененные уроки
        ### Группа
        ### Класс
        ### Ученик
        ### Учитель (учитель факультатива)
        ### Кабинет
        ### studentAll - фильтрация по занятиям для ученика, а также группам и классу, в которых он состоит
      parameters:
        - in: query
          name: start__gte
          schema:
            type: string
            format: date
            example: "2024-01-02T00:00:00+03:00"
          description: filter by great start
        - in: query
          name: end__lte
          schema:
            type: string
            format: date
            example: "2024-01-02T00:00:00+03:00"
          description: filter by less end
        - in: query
          name: number
          schema:
            type: integer
          description: filter by number of lesson
        - in: query
          name: is_cancelled
          schema:
            type: boolean
          description: filter by is_cancelled
        - in: query
          name: studentAll_id
          schema:
            type: integer
          description: filter by studentAll_id
        - in: query
          name: group_id
          schema:
            type: integer
          description: filter by group_id
        - in: query
          name: cls_id
          schema:
            type: integer
          description: filter by cls_id
        - in: query
          name: cabinet_id
          schema:
            type: integer
          description: filter by cabinet_id
        - in: query
          name: teacher_id
          schema:
            type: integer
          description: filter by teacher_id
        - in: query
          name: facultative_teacher_id
          schema:
            type: integer
          description: filter by facultative_teacher_id
        - in: query
          name: student_id
          schema:
            type: integer
          description: filter by student_id
        - in: query
          name: school_id
          schema:
            type: integer
          description: filter by school_id
        - in: query
          name: ordering
          schema:
            type: string
          description: Order by id, number, start, end, is_cancelled, lesson, group, cls, cabinet, student
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
        - in: query
          name: offset
          schema:
            type: integer
          description: The number of items to skip before starting to collect the result set
      responses:
        '200':
          content:
            application/json:
              schema: { }
          description: OK
      tags:
        - schedules

    post:
      security:
        - bearerAuth: [ ]
      operationId: CreateSchedule
      summary: Create a schedule
      description: |
        # Создание расписания уроков
        ## Cуперпользователь - создание для любых школ
        ## Активный завуч - создание только в своей активной школе
        ## В request body перечислены все возможные поля
        ### Обязательные поля: "start", "end", "lesson_id"
        ### Обязательно одно и только одно поле из "student_id", "group_id", "cls_id"
        ### Проверка на валидацию "lesson_id", "lesson_replace_id", "group_id", "cls_id", "cabinet_id", "student_id"
        ### Проверка на валидацию number - целочисленные значения от 1 до 10. !!! Указывается только в случае "шаблонного" времени уроков 
        ### Проверка cabinet_id - если указано cls_id и не указан cabinet_id, то кабинет берется класса, если не указано cls_id поле cabinet_id обязательно для заполнения
        ### Поля start и end - если создается строка с "шаблонным" временем, то в start и end на клиенте записывается шаблонное время. В этом случае поле number будет обязательным (в случае изменения шаблонного времени все будущие строки расписания будут менять значения start и end по number урока)
        ### Проверка на соответствие школ во всех внешних ключах
        ### Проверка на уникальность времени - начало урока start для ученик | группа | класс не может быть в диапазоне start-end уже существующей записи
      parameters: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                number:
                  type: integer
                  format: time
                  example: 1
                start:
                  type: string
                  format: date-time
                  example: "2000-01-01T09:50:00+03:00"
                end:
                  type: string
                  format: date-time
                  example: "2000-01-01T10:20:00+03:00"
                comment:
                  type: string
                  example: "Не забудь голову"
                replace_comment:
                  type: string
                  example: "Урок заменен по уважительной причине"
                is_cancelled:
                  type: boolean
                  example: false
                cancelled_comment:
                  type: string
                  example: "Урок отменен. Потому что так надо было"
                lesson_id:
                  type: integer
                  example: 1
                lesson_replace_id:
                  type: integer
                  example: 1
                group_id:
                  type: integer
                  example: 1
                cls_id:
                  type: integer
                  example: 1
                cabinet_id:
                  type: integer
                  example: 1
                student_id:
                  type: integer
                  example: 1
                time_lesson_id:
                  type: integer
                  example: 1
      responses:
        '201':
          content:
            application/json:
              schema: { }
          description: OK
      tags:
        - schedules
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic